refactor(agent-adapters): split approval coordinator modules
Move the approval coordinator, resolve/cancel operations, id helpers, and in-memory store out of the large approvals type file. Keep approvals.go focused on shared wire types, interfaces, constants, and sentinel errors while preserving behavior.
